<#9> feat: Album,Board,GuestNote,Setting 컨트롤러 작성

+++ src/main/java/com/j2kb/minipetpee/domain/Homepee.java
@@ -5,13 +5,14 @@ import javax.persistence.*;
 import org.hibernate.annotations.ColumnDefault;
 
 import java.util.ArrayList;
+import java.util.List;
 
 @Getter
 @Entity
 public class Homepee {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @OneToOne
     @JoinColumn(name = "member_id", nullable = false)
